+++ settings.gradle
@@ -107,8 +107,8 @@ include ':smoke-tests:apps:OpenTelemetryMetric'
 include ':smoke-tests:apps:RateLimitedSampling'
 include ':smoke-tests:apps:ReadOnly'
 include ':smoke-tests:apps:Sampling'
-include ':smoke-tests:apps:SamplingOverridesV1'
-include ':smoke-tests:apps:SamplingOverridesV2'
+include ':smoke-tests:apps:SamplingOverrides'
+include ':smoke-tests:apps:SamplingOverridesBackCompat'
 include ':smoke-tests:apps:SpringBoot'
 include ':smoke-tests:apps:SpringBootAttachInMain'
 include ':smoke-tests:apps:SpringBootAuto'
