--- smoke-tests/apps/Lettuce/src/main/java/com/microsoft/applicationinsights/smoketestapp/LettuceTestServlet.java
@@ -1,63 +1,0 @@
-/*
- * ApplicationInsights-Java
- * Copyright (c) Microsoft Corporation
- * All rights reserved.
- *
- * MIT License
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this
- * software and associated documentation files (the ""Software""), to deal in the Software
- * without restriction, including without limitation the rights to use, copy, modify, merge,
- * publish, distribute, sublicense, and/or sell copies of the Software, and to permit
- * persons to whom the Software is furnished to do so, subject to the following conditions:
- * The above copyright notice and this permission notice shall be included in all copies or
- * substantial portions of the Software.
- * THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
- * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
- * PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE
- * FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
- * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
- * DEALINGS IN THE SOFTWARE.
- */
-
-package com.microsoft.applicationinsights.smoketestapp;
-
-import io.lettuce.core.RedisClient;
-import io.lettuce.core.api.sync.RedisCommands;
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-@WebServlet("/*")
-public class LettuceTestServlet extends HttpServlet {
-
-  @Override
-  protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException {
-    try {
-      doGetInternal(req);
-      resp.getWriter().println("ok");
-    } catch (ServletException e) {
-      throw e;
-    } catch (Exception e) {
-      throw new ServletException(e);
-    }
-  }
-
-  private void doGetInternal(HttpServletRequest req) throws Exception {
-    String pathInfo = req.getPathInfo();
-    if (pathInfo.equals("/lettuce")) {
-      lettuce();
-    } else if (!pathInfo.equals("/")) {
-      throw new ServletException("Unexpected url: " + pathInfo);
-    }
-  }
-
-  private void lettuce() {
-    String hostname = System.getenv("REDIS");
-    RedisClient redisClient = RedisClient.create("redis://" + hostname);
-    RedisCommands<String, String> redisCommands = redisClient.connect().sync();
-    redisCommands.get("test");
-    redisClient.shutdown();
-  }
-}
