+++ smoke-tests/apps/SamplingOverrides/src/smokeTest/resources/applicationinsights2.json
