+++ smoke-tests/apps/SamplingOverridesBackCompat/build.gradle.kts
