+++ smoke-tests/apps/SamplingOverridesBackCompat/src/smokeTest/resources/applicationinsights2.json
