+++ smoke-tests/apps/SamplingOverridesBackCompat/src/smokeTest/resources/logback-test.xml
