--- smoke-tests/apps/SamplingOverridesV1/src/smokeTest/resources/applicationinsights.json
@@ -1,59 +1,0 @@
-{
-  "role": {
-    "name": "testrolename",
-    "instance": "testroleinstance"
-  },
-  "sampling": {
-    "percentage": 100
-  },
-  "preview": {
-    "sampling": {
-      "overrides": [
-        {
-          "spanKind": "server",
-          "attributes": [
-            {
-              "key": "http.url",
-              "value": ".*/health-check",
-              "matchType": "regexp"
-            }
-          ],
-          "percentage": 50,
-          "id": "filter out health check"
-        },
-        {
-          "spanKind": "client",
-          "attributes": [
-            {
-              "key": "db.statement",
-              "value": "select count(*) from abc",
-              "matchType": "strict"
-            }
-          ],
-          "percentage": 0,
-          "id": "filter out noisy jdbc"
-        }
-      ]
-    },
-    "instrumentationKeyOverrides": [
-      {
-        "httpPathPrefix": "/SamplingOverridesV1/login",
-        "instrumentationKey": "12345678-0000-0000-0000-0FEEDDADBEEF"
-      },
-      {
-        "httpPathPrefix": "/SamplingOverridesV1/regular-jdbc",
-        "instrumentationKey": "87654321-0000-0000-0000-0FEEDDADBEEF"
-      }
-    ],
-    "roleNameOverrides": [
-      {
-        "httpPathPrefix": "/SamplingOverridesV1/login",
-        "roleName": "app2"
-      },
-      {
-        "httpPathPrefix": "/SamplingOverridesV1/regular-jdbc",
-        "roleName": "app3"
-      }
-    ]
-  }
-}
