--- smoke-tests/apps/SamplingOverridesV2/src/smokeTest/resources/applicationinsights.json
@@ -1,63 +1,0 @@
-{
-  "role": {
-    "name": "testrolename",
-    "instance": "testroleinstance"
-  },
-  "sampling": {
-    "percentage": 100
-  },
-  "preview": {
-    "sampling": {
-      "overrides": [
-        {
-          "telemetryKind": "request",
-          "attributes": [
-            {
-              "key": "http.url",
-              "value": ".*/health-check",
-              "matchType": "regexp"
-            }
-          ],
-          "percentage": 50,
-          "id": "filter out health check"
-        },
-        {
-          "telemetryKind": "dependency",
-          "attributes": [
-            {
-              "key": "db.statement",
-              "value": "select count(*) from abc",
-              "matchType": "strict"
-            }
-          ],
-          "percentage": 0,
-          "id": "filter out noisy jdbc"
-        },
-        {
-          "telemetryKind": "trace",
-          "percentage": 100
-        }
-      ]
-    },
-    "instrumentationKeyOverrides": [
-      {
-        "httpPathPrefix": "/SamplingOverridesV2/login",
-        "instrumentationKey": "12345678-0000-0000-0000-0FEEDDADBEEF"
-      },
-      {
-        "httpPathPrefix": "/SamplingOverridesV2/regular-jdbc",
-        "instrumentationKey": "87654321-0000-0000-0000-0FEEDDADBEEF"
-      }
-    ],
-    "roleNameOverrides": [
-      {
-        "httpPathPrefix": "/SamplingOverridesV2/login",
-        "roleName": "app2"
-      },
-      {
-        "httpPathPrefix": "/SamplingOverridesV2/regular-jdbc",
-        "roleName": "app3"
-      }
-    ]
-  }
-}
